chore: refactoring style objects
